fix: update based on feedback
