+++ lib/modules/platform/bitbucket/utils.ts
@@ -1,16 +1,17 @@
 import url from 'url';
 import type { MergeStrategy } from '../../../config/types';
+import type { BranchStatus } from '../../../types';
 import { BitbucketHttp } from '../../../util/http/bitbucket';
 import type { HttpOptions, HttpResponse } from '../../../util/http/types';
 import { getPrBodyStruct } from '../pr-body';
 import type { Pr } from '../types';
-import {
+import type {
+  BitbucketBranchState,
+  BitbucketMergeStrategy,
   MergeRequestBody,
   PrResponse,
   RepoInfo,
   RepoInfoBody,
-  bitbucketMergeStrategies,
-  prStates,
 } from './types';
 
 const bitbucketHttp = new BitbucketHttp();
@@ -26,6 +27,15 @@ export function repoInfoTransformer(repoInfoBody: RepoInfoBody): RepoInfo {
   };
 }
 
+export const bitbucketMergeStrategies: Map<
+  MergeStrategy,
+  BitbucketMergeStrategy
+> = new Map([
+  ['squash', 'squash'],
+  ['merge-commit', 'merge_commit'],
+  ['fast-forward', 'fast_forward'],
+]);
+
 export function mergeBodyTransformer(
   mergeStrategy: MergeStrategy | undefined
 ): MergeRequestBody {
@@ -40,6 +50,20 @@ export function mergeBodyTransformer(
 
   return body;
 }
+
+export const prStates = {
+  open: ['OPEN'],
+  notOpen: ['MERGED', 'DECLINED', 'SUPERSEDED'],
+  merged: ['MERGED'],
+  closed: ['DECLINED', 'SUPERSEDED'],
+  all: ['OPEN', 'MERGED', 'DECLINED', 'SUPERSEDED'],
+};
+
+export const buildStates: Record<BranchStatus, BitbucketBranchState> = {
+  green: 'SUCCESSFUL',
+  red: 'FAILED',
+  yellow: 'INPROGRESS',
+};
 
 const addMaxLength = (inputUrl: string, pagelen = 100): string => {
   const { search, ...parsedUrl } = url.parse(inputUrl, true);
